remove redundant casts

+++ Patches/TaskAssignPatch.cs
@@ -262,7 +262,7 @@ internal static class RpcSetTasksPatch
         );
 
         // Convert the list of tasks to array (Il2CppStructArray)
-        taskTypeIds = new((long)TasksList.Count);
+        taskTypeIds = new(TasksList.Count);
         for (var i = 0; i < TasksList.Count; i++) taskTypeIds[i] = TasksList[i];
 
         #region Logging
